fix(APISIMP-VOCAB-BROWSE): safe long arithmetic in predicate-list page slice
Widen the `to` bound calculation to long before Math.min to prevent integer
overflow when `from + pageSize` approaches Integer.MAX_VALUE. Same overflow-safe
pattern used elsewhere (SearchV2Rest etc). Resolves CodeQL HIGH finding.

+++ backend/src/main/java/de/dlr/shepard/v2/semantic/resources/VocabularyBrowseRest.java
@@ -152,7 +152,7 @@ public class VocabularyBrowseRest {
     List<PredicateIO> all = rows.stream().map(PredicateIO::from).toList();
     int total = all.size();
     int from = (int) Math.min((long) page * pageSize, total);
-    int to = Math.min(from + pageSize, total);
+    int to = (int) Math.min((long) from + pageSize, (long) total);
     return Response.ok(new PagedResponseIO<>(all.subList(from, to), total, page, pageSize)).build();
   }
